master-ip/server/requirements.txt
# master-ip/server/requirements.txt
# Core API and server
fastapi
uvicorn[standard]

# Data validation and utilities
pydantic
python-jose[cryptography]
python-multipart

# Database (if you are using one)
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv

# QR code generation
qrcode[pil]
pydantic[email]
PyJWT
fastapi 
uvicorn 
sqlalchemy 
asyncpg 
PyJWT 
python-dotenv 
motor 
pillow
imagehash
numpy
pymongo
python-dotenv
pinecone
tqdm
torch           
transformers     
ftfy   
fastapi
uvicorn
pydantic
requests
pillow
torch
transformers
pinecone
pymongo
python-multipart         
torch
transformers
pillow
pymongo
pinecone
tqdm
sentence-transformers